Auckland Council Animal Management enforcement data

Data relating to Auckland Council Animal Management enforcement activities taken from Auckland Council Animal Management annual reports.

Usage

enforcement

Format

A dataframe with 12 rows and 9 columns:

year

Financial year data was reported

prosecutions

Total prosecutions or appeals

infringements

Total infringements issued

failure_bylaw

Infringements issued under s20(5) - Failure to comply with any bylaw authorised by the section

failure_control

Infringements issued under s53(1) - Failure to keep dog under control

failure_disqual

Infringements issued under s53(1) - Failure to comply with the effects of disqualification

failure_menacing

Infringements issued under s33EC(1) - Failure to comply with effects of classification of dog as menacing dog

failure_register

Infringements issued under s42 - Failure to register dog

obstruction_officer

Infringements issued under s18 - Wilful obstruction of dog control officer or ranger

Source

https://www.aucklandcouncil.govt.nz/en/plans-policies-bylaws-reports-projects/our-policies/policy-on-dogs.html