Auckland Council Animal Management dog impound data

Data compiled by Auckland Council's Animal Management division relating to impounded dogs in Auckland for financial years 2022-2024.

Usage

impounds

Format

Data frame with columns

docket_number

Docket number assigned to impound event.

shelter

Shelter receiving impounded animal.

team

Animal Management team responsible for impound.

impound_date

Date of impound.

impound_reason

Reason for impound.

exit_date

Date impounded animal left the shelter.

status

Reason for departure.

rfs_number

Corresponding Request for Service number.

street_name

Street animal was uplifted from.

suburb

Suburb animal was uplifted from.

s_57_type

Reason for impound, if impounded under Section 57 of the Dog Control Act.

severity

Severity of attack.

primary_breed

Primary dog breed.

secondary_breed

Secondary dog breed.

sex

Sex of the animal.

desexed

Desexed status.

approx_age

Approximate age of the animal.

dog_name

Name of dog.

dog_id

Dog ID number.

classified

Whether or not the dog has a current classification under the Dog Control Act.

current_classification

Current classification under the Dog Control Act.

classification_required

Whether or not a new classification is required as a result of circumstances leading to impounding.

new_classification

Whether or not a new classification has been issued.

new_classification_issued

New classification issued as a result of circumstances leading to impounding.

registered

Whether or not dog is registered.

new_reg_issued

Whether or not a new registration has been issued for the dog.

microchipped

Whether or not the dog is microchipped.

microchip_number

Microchip ID number.

new_microchip_issued

Whether or not a new microchip ID number has been issued.

known_dog

Whether or not the dog is 'known' to Animal Management.

euthanasia_reason

Reason for euthanasia.

days_in_shelter

Number of days in the shelter.

Source

Auckland Council, official information request (#8140017948).