Auckland Council Animal Management Request For Service (RFS) data

Data compiled by Auckland Council's Animal Management division relating to Requests For Service (Animal Management call-outs) in Auckland for financial years 2020-2025, and the second half of FY2019.

Usage

rfs

Format

Data frame with columns

sheet_name

Excel sheet from which the row originates in the raw workbook.

rfs_no

Request for Service ID number.

rfs_date

Date the RFS was lodged.

location_suburb_name

Suburb of request.

location_street_name

Street of request.

team

Animal Management team responsible for request.

local_board

Auckland Council Local Board with jurisdiction over the suburb.

rfs_type

Request type.

Source

Auckland Council, official information request (#8140017948).