Auckland Council Animal Management service data

Data relating to requests for service from Auckland Council Animal Management annual reports.

Usage

service

Format

A dataframe with 12 rows and 23 columns:

year

Financial year data was reported

p1_central

Priority 1 requests for service completed in the central region

p1_north

Priority 1 requests for service completed in the northern region

p1_south

Priority 1 requests for service completed in the southern region

p1_west

Priority 1 requests for service completed in the western region

p1_east

Priority 1 requests for service completed in the eastern region

p1_other

Priority 1 requests for service completed in other regions

p2_central

Priority 2 requests for service completed in the central region

p2_north

Priority 2 requests for service completed in the northern region

p2_south

Priority 2 requests for service completed in the southern region

p2_west

Priority 2 requests for service completed in the western region

p2_east

Priority 2 requests for service completed in the eastern region

p2_other

Priority 2 requests for service completed in other regions

agg_animals

Requests relating to dogs being aggressive to animals

agg_people

Requests relating to dogs being aggressive to people

attack_animals

Requests relating to a dog attack against an animal

attack_people

Requests relating to a dog attack against a person

barking

Requests relating to a dog barking

bylaw_breach

Requests relating to a bylaw breach

pickup_dog

Requests relating to picking up a contained dog

roaming

Requests relating to a roaming dog

unregistered_dog

Requests relating to an unregistered dog

welfare

Requests relating to a welfare check on a dog

Source

https://www.aucklandcouncil.govt.nz/en/plans-policies-bylaws-reports-projects/our-policies/policy-on-dogs.html